+++ src/main/java/com/example/d3nserver/news/controller/NewsV2Controller.java
@@ -3,7 +3,9 @@ package com.example.d3nserver.news.controller;
 import com.example.d3nserver.common.annotation.ApiDocumentResponse;
 import com.example.d3nserver.common.annotation.ReqUser;
 import com.example.d3nserver.common.dto.ResponseDto;
+import com.example.d3nserver.news.domain.Field;
 import com.example.d3nserver.news.dto.NewsResponseDto;
+import com.example.d3nserver.news.dto.TodayHomeResponseDto;
 import com.example.d3nserver.news.service.NewsService;
 import com.example.d3nserver.user.domain.User;
 import io.swagger.v3.oas.annotations.Operation;
@@ -17,6 +19,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Tag(name = "News v1.1 API", description = "News 관련 api")
 @RestController
 @RequiredArgsConstructor
@@ -25,11 +30,28 @@ public class NewsV2Controller {
     private final NewsService newsService;
 
     @ApiDocumentResponse
-    @Operation(summary = "News list", description = "Page index와 Page 크기를 받아 뉴스 페이지를 반환한다. (secondTime은 데이터가 없으면 0, selectedAnswer는 -1을 반환)")
+    @Operation(summary = "All News list", description = "Page index와 Page 크기를 받아 요청한 카테고리의 뉴스 페이지를 반환한다. (secondTime은 데이터가 없으면 0, selectedAnswer는 null을 반환)")
     @GetMapping("/list")
-    public ResponseEntity<Page<NewsResponseDto>> getAllNews(@ReqUser User user, @RequestParam @Parameter(description="페이지 인덱스, 0부터 시작")int pageIndex, @RequestParam @Parameter(description="페이지 크기")int pageSize) {
-        return ResponseDto.ok(newsService.getAllNewsDtoPageListV2(user, pageIndex, pageSize));
+    public ResponseEntity<Page<NewsResponseDto>> getAllNews(@ReqUser User user, @RequestParam @Parameter(description="페이지 인덱스, 0부터 시작")int pageIndex, @RequestParam @Parameter(description="페이지 크기")int pageSize, @RequestParam(value = "field", defaultValue = "DEFAULT") @Parameter(description = "뉴스 필드(News의 Enum 타입 Field 기입)(미기입 시 DEFAULT)")Field field) {
+        return ResponseDto.ok(newsService.getAllNewsPageList(user, pageIndex, pageSize, field));
     }
 
+    @ApiDocumentResponse
+    @Operation(summary = "Today News List", description = "type이 Recent일 경우 최신 뉴스만 반환, type이 My일 경우 맞춤형 뉴스만 반환, 지정 하지 않을 시 모두 반환")
+    @GetMapping("/today/list")
+    public ResponseEntity<List<TodayHomeResponseDto>> getTodayNews(@ReqUser User user, @RequestParam(value = "type", defaultValue = "All") String type) {
+        List<TodayHomeResponseDto> responseDtoList = new ArrayList<>();
+        if(type.equals("Recent")){
+            responseDtoList.add(new TodayHomeResponseDto("Recent",newsService.getRecentNewsList(user)));
+        }
+        else if(type.equals("My")){
+            responseDtoList.add(new TodayHomeResponseDto("My",newsService.getUserReferencedNewsList(user)));
+        }
+        else{
+            responseDtoList.add(new TodayHomeResponseDto("Recent",newsService.getRecentNewsList(user)));
+            responseDtoList.add(new TodayHomeResponseDto("My",newsService.getUserReferencedNewsList(user)));
+        }
+        return ResponseDto.ok(responseDtoList);
+    }
 
 }
