+++ src/main/java/com/example/d3nserver/news/dto/NewsResponseDto.java
@@ -1,6 +1,5 @@
 package com.example.d3nserver.news.dto;
 
-import com.example.d3nserver.common.annotation.ValidEnum;
 import com.example.d3nserver.news.domain.Field;
 import com.example.d3nserver.news.domain.News;
 import com.example.d3nserver.news.domain.NewsType;
